Beginning to merge from experimental branch some more advanced changes
Renaming some variables and functions, minor improvements to code etc.

+++ src/libcw/libcw_data.h
@@ -29,7 +29,7 @@ typedef struct cw_entry_struct{
 int          cw_representation_lookup_init_internal(const cw_entry_t *lookup[]);
 int          cw_representation_to_character_internal(const char *representation);
 __attribute__((unused)) int cw_representation_to_character_direct_internal(const char *representation);
-unsigned int cw_representation_to_hash_internal(const char *representation);
+uint8_t cw_representation_to_hash_internal(const char *representation); /* TODO: uint8_t will be enough for everyone? */
 const char  *cw_character_to_representation_internal(int c);
 const char  *cw_lookup_procedural_character_internal(int c, bool *is_usually_expanded);
